Refine OSM ingestion to avoid retaining all nodes
