+++ wildside-data/tests/features/ingest_osm_pbf.feature
@@ -26,3 +26,9 @@ Feature: ingesting OSM PBF data
     And the report lists 3 points of interest
     And the POI named "Museum Island Walk" uses the first node location
     And POIs referencing missing nodes are skipped
+
+  Scenario: ignoring irrelevant tags
+    Given a PBF file containing only irrelevant tags
+    When I ingest the PBF file
+    Then no points of interest are reported
+
