+++ wildside-data/tests/fixtures/poi_tags.osm.pbf.b64
